+++ tests/integration/flows/test_aselite_to_adls.py
@@ -1,14 +1,15 @@
+import logging
+import pandas as pd
+import os
 from typing import Any, Dict, List, Literal
 from prefect import Flow
-from viadot.tasks import AzureDataLakeUpload
+from prefect.tasks.secrets import PrefectSecret
 from prefect.run_configs import DockerRun
 from viadot.task_utils import df_to_csv, df_converts_bytes_to_int
 from viadot.tasks.aselite import ASELiteToDF
-import logging
-from viadot.flows.aselite_to_adls import ASELitetoADLS
-from prefect.tasks.secrets import PrefectSecret
-import pandas as pd
-import os
+from viadot.tasks import AzureDataLakeUpload
+from viadot.flows.aselite_to_adls import ASELiteToADLS
+
 
 TMP_FILE_NAME = "test_flow.csv"
 MAIN_DF = None
@@ -41,13 +42,8 @@ def test_aselite_to_adls():
         ,[UPD_FIELD]
     FROM [UCRMDEV].[dbo].[CRM_00]"""
 
-    RUN_CONFIG = DockerRun(
-        image="docker.pkg.github.com/dyvenia/viadot/viadot:latest",
-        labels=["prod"],
-    )
-
-    flow = ASELitetoADLS(
-        "Test flow ",
+    flow = ASELiteToADLS(
+        "Test flow",
         query=query_designer,
         sqldb_credentials_secret=credentials_secret,
         vault_name=vault_name,
@@ -61,8 +57,7 @@ def test_aselite_to_adls():
 
     MAIN_DF = pd.read_csv(TMP_FILE_NAME, delimiter="\t")
 
-    if isinstance(MAIN_DF, pd.DataFrame):
-        assert True
+    assert isinstance(MAIN_DF, pd.DataFrame) == True
 
     assert MAIN_DF.shape == (10, 17)