+++ viadot/tasks/__init__.py
@@ -25,6 +25,7 @@ from .sqlite import SQLiteInsert, SQLiteSQLtoDF, SQLiteQuery
 from .supermetrics import SupermetricsToCSV, SupermetricsToDF
 from .sharepoint import SharepointToDF
 from .cloud_for_customers import C4CReportToDF, C4CToDF
+from .aselite import ASELiteToDF
 
 try:
     from .sap_rfc import SAPRFCToDF